Update README and example with new screenshots and links
This commit updates the README.md file with new screenshots showcasing the package's functionality. It also replaces the existing screenshots in the `screenshots` directory with new ones: `gif.gif`, `help.png`, and `home.png`.

Additionally, the example in `example/lib/main.dart` has been updated with a new Tawk.to direct chat link and includes a comment prompting users to replace it with their own. The `onAgentMessage` callback in the example now also includes a comment suggesting actions to take with the agent's message.