This commit modifies the `WebViewWidget` configuration in `lib/src/tawk_service.dart`.
The `transparentBackground` property of `WebWebViewWidgetCreationParams` is now set to `true`, allowing the web content to have a transparent background if supported.

The `useHybridComposition` property is set to `false`, which may impact rendering performance or behavior depending on the platform and web content.

Additionally, the delay before hiding the placeholder in `onLoadStop` has been reduced from 2 seconds to 1 second.